Add load method to the storage, refactor some types from float to double

+++ src/main/java/chching/record/Record.java
@@ -4,19 +4,12 @@ public abstract class Record {
     protected String category;
     protected String description;
     protected String date;
-    protected float value;
+    protected double value;
 
-    /**
-     * Generic constructor for Record class.
-     * @param category string of record category, accepts "income" or "expense"
-     * @param description string of record description
-     * @param date string of record date
-     * @param value float of record amount
-     */
     public Record(){
 
     }
-    public Record(String category, String description, String date, float value) {
+    public Record(String category, String description, String date, double value) {
         this.category = category;
         this.description = description;
         this.date = date;
@@ -36,7 +29,7 @@ public abstract class Record {
         return date;
     }
 
-    public float getValue() {
+    public double getValue() {
         return value;
     }
